refactor/reorganize matrix.cc
The original motivation was to use cudaMallocPitch. The minimum row-size seems to be
512 bytes on my machine, and that way nothing fits into my memory, because CIFAR-10
involves many short matrix rows. Even when I make it fit, it seems to be slower,
most likely due to the vector-like (map) operations. Thus, I kept the cleanup-like
parts of the change, but reverted the rest.